refine mcp tools
